Add footnotes support to markdown
