+++ src/lib/utils.ts
@@ -1,10 +1,6 @@
-import type { marked } from "marked";
-
-import katex from "katex";
 import md5 from "md5";
 import twemojiModule from "twemoji";
 
-import emojis from "@app/lib/emojis";
 import { assert } from "@app/lib/error";
 import { base } from "@app/lib/router";
 import { config } from "@app/lib/config";
@@ -208,14 +204,6 @@ export function getDaysPassed(from: Date, to: Date): number {
   return Math.floor((to.getTime() - from.getTime()) / (24 * 60 * 60 * 1000));
 }
 
-export function parseEmoji(input: string): string {
-  if (input in emojis) {
-    return emojis[input];
-  }
-
-  return input;
-}
-
 export function scrollIntoView(id: string) {
   const lineElement = document.getElementById(id);
   if (lineElement) lineElement.scrollIntoView();
@@ -264,91 +252,19 @@ export const unreachable = (value: never): never => {
   throw new Error(`Unreachable code: ${value}`);
 };
 
-const emojisMarkedExtension = {
-  name: "emoji",
-  level: "inline",
-  start: (src: string) => src.indexOf(":"),
-  tokenizer(src: string) {
-    const match = src.match(/^:([\w+-]+):/);
-    if (match) {
-      return {
-        type: "emoji",
-        raw: match[0],
-        text: match[1].trim(),
-      };
-    }
-  },
-  renderer: (token: marked.Tokens.Generic) =>
-    `<span>${parseEmoji(token.text)}</span>`,
-};
-
-const katexMarkedExtension = {
-  name: "katex",
-  level: "inline",
-  start: (src: string) => src.indexOf("$"),
-  tokenizer(src: string) {
-    const match = src.match(/^\$+([^$\n]+?)\$+/);
-    if (match) {
-      return {
-        type: "katex",
-        raw: match[0],
-        text: match[1].trim(),
-      };
-    }
-  },
-  renderer: (token: marked.Tokens.Generic) =>
-    katex.renderToString(token.text, {
-      throwOnError: false,
-    }),
-};
-
-// Converts self closing anchor tags into empty anchor tags, to avoid erratic wrapping behaviour
-// e.g. <a name="test"/> -> <a name="test"></a>
-const anchorMarkedExtension = {
-  name: "sanitizedAnchor",
-  level: "block",
-  start: (src: string) => src.match(/<a name="([\w]+)"\/>/)?.index,
-  tokenizer(src: string) {
-    const match = src.match(/^<a name="([\w]+)"\/>/);
-    if (match) {
-      return {
-        type: "sanitizedAnchor",
-        raw: match[0],
-        text: match[1].trim(),
-      };
-    }
-  },
-  renderer: (token: marked.Tokens.Generic) => {
-    return `<a name="${token.text}"></a>`;
-  },
-};
-
-// Overwrites the rendering of heading tokens.
-// Since there are possible non ASCII characters in headings,
-// we escape them by replacing them with dashes and,
-// trim eventual dashes on each side of the string.
-export const renderer = {
-  heading(text: string, level: 1 | 2 | 3 | 4 | 5 | 6) {
-    const escapedText = text
-      .toLowerCase()
-      .replace(/[^\w]+/g, "-")
-      .replace(/^-|-$/g, "");
-
-    return `<h${level} id="${escapedText}">${text}</h${level}>`;
-  },
-};
-
-export function twemoji(node: HTMLElement) {
+export function twemoji(
+  node: HTMLElement,
+  { exclude }: { exclude: string[] } = { exclude: [] },
+) {
   twemojiModule.parse(node, {
+    callback: (icon, options: Record<string, any>) => {
+      return exclude.includes(icon)
+        ? false
+        : "".concat(options.base, options.size, "/", icon, options.ext);
+    },
     base,
     folder: "twemoji",
     ext: ".svg",
     className: `txt-emoji`,
   });
 }
-
-export const markdownExtensions = [
-  emojisMarkedExtension,
-  katexMarkedExtension,
-  anchorMarkedExtension,
-];
